pyproject.toml
[project]
name = "db_update"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.1",
    "msgspec>=0.19.0",
    "polars>=1.20.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "aiofiles>=24.1.0",
    "asyncpg-stubs>=0.30.0",
    "nox>=2025.5.1",
    "pyright>=1.1.391",
    "ruff>=0.8.4",
]
lint = ["ruff>=0.11.9"]
typecheck = ["pyright>=1.1.400"]
format = ["ruff>=0.11.9"]
